Users/lionh/Downloads/invoice_extractor_raw/requirements.txt
# Core ML and AI Libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0

# Computer Vision and Image Processing
Pillow>=9.5.0
opencv-python>=4.8.0
PyMuPDF>=1.23.0  # For PDF processing

# Web UI Framework
streamlit>=1.25.0

# Data Processing and Utilities
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# File and System Operations
pathlib2>=2.3.7; python_version < "3.4"
shutil-which>=1.0.0

# JSON and Data Serialization
json5>=0.9.14

# Regular Expressions (built-in, but ensuring compatibility)
regex>=2023.6.0

# Progress Bars and Logging
tqdm>=4.65.0
rich>=13.4.0

# Optional: GPU Support (uncomment if using CUDA)
# torch-cuda>=2.0.0

# Development and Testing (optional)
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0

# Jupyter Notebook Support (optional)
jupyter>=1.0.0
ipykernel>=6.25.0

# Additional Utilities
python-dateutil>=2.8.2
requests>=2.31.0
urllib3>=2.0.0

# For better performance on Windows
pywin32>=306; sys_platform == "win32" 